+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [6.2.2] - 2024-04-29
+### Fixed
+* Fixed bug in resource controller when model set in the method.
+### Added
+* Factory running if exists in the model.
+
 ## [6.2.1] - 2024-04-26
 
 ### Fixed